Rename next-edge -> fetch-api
