+++ src/handlers/fetch-api/create-manifest-handler.ts
