+++ src/handlers/shared/validate-allow-saleor-urls.ts
